Solves bug in CommentAutoIndentStrategy
Because of the new introduced Scaladoc code blocks the end of open
Scaladoc comments was not correctly detected anymore. Before, there was
looked if the actual partition reaches the end of the file, now it is
looked if the last partition of the file is a Scaladoc partition. This
makes a difference because Scaladoc comments are divided into multiple
parts when they contain code blocks, thus the part containing the start
of the Scaladoc section is not the same as the part containing the end
of the Scaladoc section.

Fix #1001463