Correct behavior of multi line comments before string literals
This is not a full fix for the problem that multi line comments are not
automatically closed when they are followed by a string containing the
*/ sign. It is difficult to completely solve this problem because
whenever one opens a multi line comment with /* the rest of the document
does no longer contain any partitions because the comment spans the
complete content. Thus, it is not possible anymore to detect if the
closing sign */ is located inside of strings or not.

The behavior introduced solves the problem as good as possible. What it
does is it looks if the line containing a */ sign contains a quotation
mark. If this is the case it is assumed that the ending is located
inside of a string literal, thus the multi line comment is closed.

This fix doesn't work if the */ sign is located inside of multi line
string literals or if the string literals aren't even closed but it
helps at least partially.

Fixes #1001515