``ScalaPresentationReconciler`` doesn't need to inherit from ``JavaPresentationReconciler``
Alright, the change was so simple that I could not stopping myself from doing it. I honestly
didn't realize we had a ``ScalaClassFileEditor``. This simple fact makes my big comment about
not being able to prevent Java semantic highligting to be enabled on Scala source attachments
not true. Indeed, it's straightforward to do so!