Revert "Merge pull request #269 from mads379/parsetree-1001326"
The parse-tree changes have been merged in Scala [pull 1982](https://github.com/scala/scala/pull/1982),
so there is no need for it anymore. Moreover, the presentation compiler must *not* be interrupted
while parsing on a foreign thread, and that can't be ensured in this commit (but is correctly ensured
in the Scala changeset). Since this branch is used with 2.9.3, there is a potential for race conditions.

This reverts commit 712e6bf5eb983a18aac03294c0656519d53ac634, reversing
changes made to 4f7b391cf1a64c6d906263afc2ebd0928bf4cb4d.