Adds multi line error messages
The new behavior is more or less the behavior that is inherited from
JDT. The only difference is that error messages aren't shrank to
single lines anymore but are surrounded with a HTML-pre-code block. This
allows to display them in the Annotation Hover as well as in the
Problems View on multiple lines.

Fixes #1001534