Fixed ``SemanticHighlightingPositionsTest``
With the new reconciliation strategy implementation, compilation units are not
reloaded by the presentation compiler at every buffer change. Hence, in the
semantic highlighting positions tests, we need to manually force the reload of
the unit after its content is changed, or the tests will fail because the
semantic highlighting reconciler would be performed on the old unit's content.

Reload the unit after its content is replaced