Corrections and improvements from code reviews
The name has been changed to "Identifier in interpolated string" since now all
identifiers, including methods, are styled. This is important because writing
s"$method" is valid and should be styled.

We also need to listen for property change events which happen when a user
changes their preferences, so that the editor can be updated.