Correct syntax coloring for //* in multi line comments
Whenever two consecutive slashes occur before a star inside of a multi
line comment the rule is that such a comment is not terminated. This
means /*//*/ is a valid comment. On the other hand /*/*//*/*/ is not a
valid comment because whenever the //* occurs inside of a nested
comment then it is counted as a starting point of another nested
comment.

This behavior looks weird but it reflects the way how the compiler
works.

Fixes #1000790