Add semantic highlighting for type Dynamic
Calls like `d.foo(1)` or `d.foo(param = 1)` are treated as method
calls whereas calls like `d.foo` or `d.foo = 1` are treated as vars.

It can be that a call to `d.foo` must not be a mutable field but an
immutable one when no corresponding call to `updateDynamic` is done
before. However, in such a case `foo` is highlighted as a var becuase
it is impossible for the IDE to detect such runtime informations.

Fixes #1001656