Updated Sbt version used by the different Scala profiles
* When building for Scala 2.10.x, let's use Sbt 0.13.0-M1 artifacts
* When building for Scala 2.11.x, let's use the Sbt 0.13.0 SNAPSHOT artifacts

There is an additional catch. When building for Scala 2.10.x, the value of the
property ``sbt.version`` has a trailing ``SNAPSHOT``. Why? Becuase we are using
Scala 2.10.2-SNAPSHOT, and hence the Sbt artifacts have to be rebuilt **every
night**. In the Maven world this means that the artifacts version has to end with
``SNAPSHOT`` (there is really no way around this).

In case you decide to build for Scala 2.10.0 or Scala 2.10.1, then make sure to
pass ``-Dsbt.version=0.13.0-M1`` to the build script (no SNAPSHOT is needed
in this case because the Scala version is stable!).