Remove dependency on GenASM for structure builder.
We relied on genJVM, and later on genASM, to retrieve method descriptors for
method parameter types in the structure builder. That part of the code
proved to be a bit unstable, and there is no reason why we should continue to
rely on it.

This commit implements the same functionality on our side.

* Sorry for the whitespace changes in `ScalaJavaMapper`. The only meaningful changes
are in `mapParamTypeSignature`.
* the current code still relies on scalac backend, just one step earlier (icode).