Remove dependency on GenASM for structure builder.
We relied on genJVM, and later on genASM, to retrieve method descriptors for
method parameter types in the structure builder. That part of the code
proved to be a bit unstable, and there is no reason why we should continue to
rely on it.

This commit implements the same functionality on our side.

* the current code still relies on scalac backend, just one step earlier (icode).
* This is a non-whitespace-changing rework of https://github.com/scala-ide/scala-ide/pull/381

(cherry picked from commit e0d5d6dcd53598fd3f939fab9747507247578137)

Conflicts:
	org.scala-ide.sdt.core/src/scala/tools/eclipse/JVMUtils.scala