Revert "Do not lock the whole workspace when running the ``ProjectsCleanJob``"
When performing a project's clean (i.e., via
`project.build(IncrementalProjectBuilder.CLEAN_BUILD, monitor)), the whole
workspace gets lock. Therefore, the `ProjectsCleanJob` needs to make sure to set
a rule to lock the whole workspace when creating the `Job`, or an
`IllegalArgumentException` will be thrown during the call to `project.build`.

It turns out that the call to `ResourcesPlugin.getWorkspace().getRuleFactory().buildRule()`
actually does the right thing, i.e., it creates a rule that locks the whole workspace.

This reverts commit 24ec637983d4afa24de23f91edcd353b6204bf35.

Fixed #1001658