Deprecation warnings no longer accumulate
The reason why that used to happen is that the Eclipse markers attached to the
dependent sources recompiled by Sbt were not removed, and hence the deprecation
warnings (or any other kind of problem marker) were being accumulated after
each incremental build.

The fix consists in deleting all problem's marker attached to a unit when the
unit is about to be parsed. This should really be seen as a workaround until
Sbt provides a better interface for incremental compilation. In fact, ideally
Sbt should provide a notification mechanism to know what is the set of
compilation units that are about to be recompiled, **before** the actual
recompilation step is performed.  Unfortunately, that is not possible at the
moment, hence we are left with the current hack.

Huge thanks to @gkossakowski for suggesting this very effective workaround.

Fixed #1001595