Remove "private" compiler options from the settings dialog.
I don't think anyone ever used these settings from the IDE,
and they can be accessed using the "additional params" edit
(with completion and validation). Moreover, we had an arbitrary
subset of the tens of `-Y` parameters.

These parameters are subject to change, and the 2.11 build
fails because one of them was removed. This fixes the nightly
build for 2.11, in the wake of 2.11.0-M3.