Use the Scala IDE nightly update-site to build the codebase
This commit simplifies the process of building the Scala IDE codebase.

First, the toolchain no longer need to be built. Why? Because the Scala IDE
dependencies (scala, scalariform, sbt, scala-refactoring) are now retrieved
from the Scala IDE nightly update-site, which of course contains the needed
dependencies.  Of course, the toolchain will still be needed for building the
Scala IDE nightly, but this only need to be done once and the uber-script can
make sure that all dependencies are built with the same Scala version (this
point is extremely important when building against a Scala version such as
2.11.0-SNAPSHOT, because of potential binary incompatibilities, both in the
library and the compiler).

Second, I've been doing some clean-up in the POMs. This is mostly related to
the fact that we no longer support Scala 2.9.

Third, I moved some of the properties/dependencies/plugin configuration from
the root POM, down one level. The rationale is that when looking at the Scala
IDE build file you shouldn't need to know about configuration values that are
specific to building the toolchain. (actually, I need to do more work to bring
this to completion, but I wanted to first check you guys aren't against doing
so).