Exposed "Restart Presentation Compiler" action
Users can now restart the presentation compiler at will.  This can be handy
when false errors are reported in a compilation unit, and you want to reset a
Presentation Compiler without performing a full clean/build.

To manually restart a Presentation Compiler, a user has to right click on the
project's folder (in the Package Explorer), and select ``Scala > Restart
Presentation Compiler``.

In order to place the functionality inside the Scala menu (as described above),
I was forced to use an `action` (and not a `command`). I did try to use a
`command`, but I couldn't find the incantantion for placing the "Restart
Presentation Compiler" inside the Scala menu when using a `menuContribution`.

Fixes #1000555