Removes references to JavaStringAutoIndentStrategy
This is an internal JDT class, and its constructor changed between Juno and Kepler.
It was a blocker to be able to use the same build on both Juno and Kepler.
Less non-API reference is better.

The features lost are the one defined in the 'In string literals' in
Preferences>Java>Edito>Typing. They are about managing multiline content in simple string literals.
They didn't really make in the Scala context, where multiline string literals can be used.