Fix create method quickfix exception
The create method quickfix failed with an exception (calling None.get) when the error message triggering the quickfix was "not found: value <method name>".
This is because we would mistakenly tell the refactoring library that the target was always a class, even though it could have been something else (eg, an object).
The scala-refactoring library now supports targetting the definition by offset, so we use that here.

Fix #1001740