Rename Colour to Color to improve compatibility...
... with third-party plugins like the Eclipse Color Theme Plugin

The important part is the change in
  val FOREGROUND_COLOUR_SUFFIX = ".colour"
but considering that Eclipse itself seems to have converged on “Color”,
I renamed everything appropriately to not create confusion whether color
or colour needs to be used.

With this change the Eclipse Color Theme Plugin “just works” (one commit
against Eclipse Color Theme Plugin pending, which updates the mapping).