Resolves a bug that leads to no quick assistants
This was an interesting bug to fix. What happened here was that we
had a parallel collection that swallowed an exception. Because the
exception was never shown to the outside world it was not possible
to find out what went wrong. I managed to get the exception by
removing the parallel collection, which was done not because I knew
what to do but because the parallel collection looked unnecessary.

The exception itself was due to an unescaped string that is interpreted
as a regex and could easily be fixed. It occurred whenever one selected
a method with a symbolic identifier like '+'.

More interestingly, even after the fix of the bug the code with the
parallel collection still didn't work - it just returned an empty list.
I have no idea what went wrong but I believe its a problem somewhere in
the internals of the parallel collection - thus no usege of it anymore.

Fixes #1001801, #1001798