Resolve exception when using type mismatch quick fix
This is a regression due to refactorings in 8f42b9de31c9365d6f9c.

The regex that was used to check an internal message was updated to
work with Scalas new string interpolation feature. Unfortunately,
while the regex was the same the semantics changed because it was
tried to insert and extract a value into and from the string
interpolator at the same time, which is in this situation not possible.

Fixes #1001809