Two new profiles for supporting legacy and new sbt dependencies
This commit allows us to move our build to consume the sbt artifacts created
and published by dbuild using [the following
configs](https://github.com/typesafehub/sbt-builds-for-ide), without breaking
the [Scala/IDE integration
script](https://github.com/scala/jenkins-scripts/blob/master/job/pr-scala-integrate-ide),
i.e., we maintain backward compatibility.

This of course is only meant to be an interim solution, in fact we are already working
on updating the Scala/IDE integration script to be compatible with the new build. However,
it will take some time before we can ship the new integration script, hence the idea of
supporting both way of building for a short amount of time.

One inconvenient of doing so is that a new profile must be now passed to the build.

* sbt-new: This is the profile defining the new sbt dependencies. In general, you will
           always want to run the build with this profile enabled.

* sbt-legacy: This is the profile allowing compatibility with the Scala/IDE integration
              script. There is really no reason for you to use this, unless you know what
              you are doing.

Once the transition to the new sbt dependencies is performed, we can revert this commit.

Credit for this commit goes to @huitseeker