Do not export sbinary and jline packages in ``org.scala-ide.sbt.full.library``
We don't seem to need to export the sbinary and jline packages. Moreover, we
used to export these packages with the **wrong** version number, i.e., we were
using the ${sbt.version}, instead of the actual version the sbt build file
declared for these dependencies. Hence, I think we are better off not exporting
these packages anymore in the MANIFEST produced for
``org.scala-ide.sbt.full.library`` bundle.