Fixed conditional breakpoint in loop bug.
I was noticing that conditional break points in loops were only being evaluated every other iteration of the loop. Turns out that if you don't want to break at a conditional breakpoint, in BreakpointSupportActor's BreakpointEvent behaviour, all that is necessary is to reply false. No need to also resume the thread as I was doing before!