Uses new debug view selection support
In Eclipse 4.x, a new way to follow the current selection in the debug view was implemented,
as debug specific support.
The old way was slightly broken, and was no more working for selection changes not triggered
by the user. A temporary fix was implemented in Scala 3.0.x, to support Eclipse 3.x and 4.x.
This commit replace the old code and the fix by a solution using the new support.

Fixes #1001585
(cherry picked from commit 22238324758784a8bc9cdb1db6ce5f134b3420c3)